QuabStream.cc was separated from QuabResource.cc
